Moved non-lines to EditboxOther
